pyproject.toml
[tool.poetry]
name = "tt-nmx"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "tt_nmx"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
psycopg2-binary = "^2.9.7"
django-environ = "^0.11.2"
django-filter = "^23.2"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
